requirements.txt
gspread==5.1.1
oauth2client==4.1.3
discord.py==2.1.0